Add theme system and Prettier to the project
This commit includes the introduction of a theme system to the project, structured into separate files for base settings, breakpoints, mixins, palette, transitions, typography, shadows, zIndex, and theme specific overrides (dark and light). In addition to this, Prettier was added for code formatting and enforced by a new 'prettier-check' script.